split card contacts
