+++ libs/ui/elements/src/lib/internal-link-card/internal-link-card.component.html
@@ -71,80 +71,10 @@
           </div>
         </div>
 
-        <div
-          data-cy="recordOrg"
-          *ngIf="size !== 'XS' && record.ownerOrganization?.name"
-          class="flex items-center justify-evenly bg-gray-50 rounded-lg h-[53px] px-2"
-        >
-          <div class="flex items-center flex-1 min-w-0">
-            <div
-              class="w-[45px] h-[45px] rounded-lg overflow-hidden shrink-0 mr-3"
-            >
-              <gn-ui-thumbnail
-                [thumbnailUrl]="
-                  record.ownerOrganization?.logoUrl?.toString() || ''
-                "
-                [fit]="'contain'"
-                class="w-full h-full rounded-lg"
-              ></gn-ui-thumbnail>
-            </div>
-            <div *ngIf="organization?.name" class="flex-1 w-0 overflow-hidden">
-              <div
-                class="text-xs text-black font-normal leading-tight truncate"
-                translate
-              >
-                record.card.metadata.contact
-              </div>
-              <div
-                data-cy="recordOrgName"
-                class="text-xl text-primary-black font-medium truncate"
-              >
-                {{ organization.name }}
-              </div>
-            </div>
-          </div>
-          <div *ngIf="size === 'L'" class="ml-2 flex space-x-2">
-            <div *ngIf="organization?.website" class="flex">
-              <button
-                [title]="organization.website"
-                class="w-[40px] h-[32px] flex items-center justify-center rounded-lg border border-[#D4D3D7] px-[8px] py-[4px] hover:bg-primary-lightest"
-                (click)="openExternalUrl($event, organization.website)"
-              >
-                <ng-icon name="iconoirInternet"></ng-icon>
-              </button>
-            </div>
-            <div *ngIf="contacts[0]?.email" class="flex">
-              <button
-                [title]="contacts[0].email"
-                class="w-[40px] h-[32px] flex items-center justify-center rounded-lg border border-[#D4D3D7] px-[8px] py-[4px] hover:bg-primary-lightest"
-                data-cy="contact-email"
-                (click)="openMailto($event, contacts[0].email)"
-              >
-                <ng-icon name="matEmailOutline"></ng-icon>
-              </button>
-            </div>
-            <div *ngIf="contacts[0]?.phone" class="flex">
-              <button
-                [title]="'Copy to clipboard'"
-                class="w-[40px] h-[32px] flex items-center justify-center rounded-lg border border-[#D4D3D7] px-[8px] py-[4px] hover:bg-primary-lightest relative group"
-                data-cy="contact-phone"
-                (click)="copyToClipboard($event, contacts[0].phone)"
-              >
-                <ng-icon name="matPhoneOutline"></ng-icon>
-              </button>
-            </div>
-            <div *ngIf="contacts[0]?.address" class="flex">
-              <button
-                [title]="'Copy to clipboard'"
-                class="w-[40px] h-[32px] flex items-center justify-center rounded-lg border border-[#D4D3D7] px-[8px] py-[4px] hover:bg-primary-lightest relative group"
-                data-cy="contact-phone"
-                (click)="copyToClipboard($event, contacts[0].address)"
-              >
-                <ng-icon name="matLocationOnOutline"></ng-icon>
-              </button>
-            </div>
-          </div>
-        </div>
+        <gn-ui-internal-link-card-contact
+          [record]="record"
+          [size]="size"
+        ></gn-ui-internal-link-card-contact>
       </div>
     </div>
   </div>
